Fix flakey test for florida river (#163)
Fixes the test which was previously failing, saying that the Florida
River wasn't in the test graph.

It appears the URIs were changed and as a result, the geoconnex.us ones
got switched to `pids.geoconnex.dev `

@webb-ben could you confirm if this fix to the test seems reasonable? My
one apprehension is that it seems like we have some URIs that are
pids.geoconnex.dev but some there are left over for the linked mainstems
which are `geoconnex.us`. I am not sure if we want to have all the URIs
use the same basename; happy to defer to whatever you think is best.

i.e. this query still works since it appears the jsonld properties don't
refer to `pids.geoconnex.us` but rather the prod ones.

```sparql
    PREFIX hyf: <https://www.opengis.net/def/schema/hy_features/hyf/>

    select * where { 
        ?monitoringLocation hyf:referencedPosition/hyf:HY_IndirectPosition/hyf:linearElement <https://geoconnex.us/ref/mainstems/42750> .
    } limit 100 
```