Add flatgeobuf download for metadata (#188)
In order to maximize the speed of mainstem resolution it is best to have
the mainstem file locally. This file is around 6 gigabytes. Thus I have
an asset for this in the config group but it is skipped in both testing
and ci/cd since it is relatively trivial and just downloads a file